requirements.txt
# Flask Web Framework
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.0
blinker==1.9.0

# Cloudinary for image hosting
cloudinary==1.36.0

# Production server
gunicorn==21.2.0

# Additional utilities
python-dotenv==1.0.0  # For environment variables

# Supabase Python Client
supabase==2.11.0